% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_to_all.R
\name{add_to_all}
\alias{add_to_all}
\title{Will add a question to the data.frame of all current questions}
\usage{
add_to_all(x, addto)
}
\arguments{
\item{x}{The questions to be added to the data.frame}

\item{addto}{The data.frame the question should be added to}
}
\value{
The new data.frame with the new question added
}
\description{
Will add a question to the data.frame of all current questions
}
