% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_to_all.R
\name{add_to_all_test}
\alias{add_to_all_test}
\title{Will add a test to the list of all current tests}
\usage{
add_to_all_test(x, addto)
}
\arguments{
\item{x}{The list containing the test name and the name of all questions in the test}

\item{addto}{The list of tests to be added to}
}
\value{
The list of tests containing the new test
}
\description{
Will add a test to the list of all current tests
}
