% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/answer_question&mark_question.R
\name{answer_question}
\alias{answer_question}
\title{Given a student and a question, gives a random answer to the question from the student}
\usage{
answer_question(s, q)
}
\arguments{
\item{s}{A student that will answer the question}

\item{q}{A question that the student will answer}
}
\value{
A new student with the updated marks from the student answering the question
}
\description{
Given a student and a question, gives a random answer to the question from the student
}
