% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete.R
\name{delete}
\alias{delete}
\title{Will delete a question or a test from the collection of all questions or tests}
\usage{
delete(x, removefrom)
}
\arguments{
\item{x}{The name of the question or test that should be deleted}

\item{removefrom}{The collection of either questions or tests that the items should be deleted}
}
\value{
The collection of questions or tests with the named item deleted
}
\description{
Will delete a question or a test from the collection of all questions or tests
}
