% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_question}
\alias{extract_question}
\title{Will get a question from a list of all currently stored questions via its name}
\usage{
extract_question(name, all_qs)
}
\arguments{
\item{name}{The name of the desired question}

\item{all_qs}{A data.frame of all currently saved questions}
}
\value{
The question with the given name if such a question exists
}
\description{
Will get a question from a list of all currently stored questions via its name
}
