% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_test}
\alias{extract_test}
\title{Will return a test object given the name of the test and a list of all tests and all questions}
\usage{
extract_test(name = character, all_ts = list(), all_qs = data.frame())
}
\arguments{
\item{name}{The name of the desired test}

\item{all_ts}{The list of all currently stored tests}

\item{all_qs}{A data.frame of all the currently stored questions}
}
\value{
The test with the given name if it exists
}
\description{
Will return a test object given the name of the test and a list of all tests and all questions
}
