% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_test.R
\name{generate_test}
\alias{generate_test}
\title{Will Generate a test based on a list of question names}
\usage{
generate_test(all_qs = data.frame(), qs = character(), can_neg = logical())
}
\arguments{
\item{all_qs}{A data.frame containing all current created questions}

\item{qs}{A list of all the question names you wish to add to the test}

\item{can_neg}{A logical argument. TRUE will mean a student can achieve a negative score on the test}
}
\value{
A new test containing the listed questions
}
\description{
Will Generate a test based on a list of question names
}
