% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_percentage.R
\name{get_percentage}
\alias{get_percentage}
\title{Will get the percentage of marks scoring above a specified percentage}
\usage{
get_percentage(marks = numeric(), percentage = integer(), max_mark = numeric())
}
\arguments{
\item{marks}{A vector containing the marks of all the students}

\item{percentage}{The percentage we are interested in}

\item{max_mark}{The maximum possible mark on the test}
}
\value{
The percentage of marks above the specified percentage
}
\description{
Will get the percentage of marks scoring above a specified percentage
}
