% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_question_names.R
\name{get_question_names}
\alias{get_question_names}
\title{Will get the names of each of the questions in a test}
\usage{
get_question_names(t = test())
}
\arguments{
\item{t}{The test that the question names should be extracted from}
}
\value{
A vector of all the question names from the test
}
\description{
Will get the names of each of the questions in a test
}
