% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCI.R
\name{getPerCI}
\alias{getPerCI}
\title{Will get the confidence interval for the percentage of students achieving a pass mark}
\usage{
getPerCI(marks, percentage, max_mark)
}
\arguments{
\item{marks}{A vector containing the marks of all the students}

\item{percentage}{The percentage we are interested in}

\item{max_mark}{The maximum possible mark on the test}
}
\value{
The confidence interval for the percentage of students that will pass
}
\description{
Will get the confidence interval for the percentage of students achieving a pass mark
}
