% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_test.R
\name{in_test}
\alias{in_test}
\title{Will get a list of all tests that a question is contained in}
\usage{
in_test(x, all_tests)
}
\arguments{
\item{x}{The name of the question}

\item{all_tests}{The list of all currently saved tests}
}
\value{
A list of the names of tests that the question is contained in
}
\description{
Will get a list of all tests that a question is contained in
}
