% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_all.R
\name{load_all_questions}
\alias{load_all_questions}
\title{Will load all the questions currently saved in 'questions.csv'}
\usage{
load_all_questions()
}
\value{
A data.frame of all the currently saved questions
}
\description{
Will load all the questions currently saved in 'questions.csv'
}
