% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_all.R
\name{load_all_tests}
\alias{load_all_tests}
\title{Will load the names of all the tests currently stored in 'tests.csv'}
\usage{
load_all_tests()
}
\value{
A list of the names of all the currently saved tests
}
\description{
Will load the names of all the tests currently stored in 'tests.csv'
}
