% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/answer_question&mark_question.R
\name{mark_question}
\alias{mark_question}
\title{Mark a question that a student has answered}
\usage{
mark_question(s = student(), q = question())
}
\arguments{
\item{s}{A student that has an answer to the question given}

\item{q}{A question that is being marked witht the students answers}
}
\value{
A student with the marks updated based on if the answer was correct
}
\description{
Mark a question that a student has answered
}
