% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_student.R
\name{new_student}
\alias{new_student}
\title{Creates a new object with class 'student'}
\usage{
new_student(answer = integer(), marks = numeric())
}
\arguments{
\item{answer}{A list of option numbers that the student will use as the answer to the question}

\item{marks}{The current amount of marks that the student has}
}
\value{
A new Student class object
}
\description{
Creates a new object with class 'student'
}
