% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_test.R
\name{new_test}
\alias{new_test}
\title{Creates a new object with class 'test'}
\usage{
new_test(questions = list(), name = chr(), can_neg = logical())
}
\arguments{
\item{questions}{A list of questions that makes up the test}

\item{name}{The name of the test}

\item{can_neg}{A logical value. set to TRUE if a student should be able to achieve negative marks taking the test}
}
\value{
A new Test class object
}
\description{
Creates a new object with class 'test'
}
