% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_item.R
\name{save_item}
\alias{save_item}
\title{Will save the data.frame of questions or a list of tests to a .csv file}
\usage{
save_item(x)
}
\arguments{
\item{x}{The item to be saved}
}
\value{
NA
}
\description{
Will save the data.frame of questions or a list of tests to a .csv file
}
