% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/take_test.R
\name{take_test}
\alias{take_test}
\title{Makes a given student answer all the questions in a given test}
\usage{
take_test(s, t)
}
\arguments{
\item{s}{The student that will take the test}

\item{t}{The test that the student will answer}
}
\value{
A student with updated marks reflecting the results of answering the test
}
\description{
Makes a given student answer all the questions in a given test
}
